M2U-Net/requirements.txt
numpy==1.21.6
onnx==1.10.2
Pillow==9.0.0


